% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rsi.R
\name{add_rsi}
\alias{add_rsi}
\title{Add the Relative Strength Index (RSI) to a data set of prices}
\usage{
add_rsi(data, time_period = "1 hour", n = 7, cutoff = 30)
}
\arguments{
\item{data}{A \code{data.table} containing \code{time} and \code{price} columns.}

\item{time_period}{A character specifying the time frame for calculating
the RSI. Possible values "3 minutes", "5 minutes", "15 minutes",
"30 minutes", "1 hour", "2 hours", "4 hours", "8 hours", "1 day". Default
is "1 hour".}

\item{n}{Number of periods for moving averages.}

\item{cutoff}{A cutoff value for creating a logical vector "deal_start". By
default, a cutoff value of 30 is used, which means that "deal_start" is TRUE
when the RSI is below 30 and FALSE otherwise.}
}
\value{
A \code{data.table} with added RSI and logical vector \code{deal_start}.
}
\description{
This function takes a data set of price data and adds a column containing the
RSI and a column containing a logical vector that is TRUE when the RSI is
below the specified cutoff and FALSE otherwise. It can be used to prepare
the data for using the \code{\link[=bot]{bot()}} function with \code{start_asap = FALSE}.
}
\examples{
#Download some price data
dat <- get_binance_prices_from_csv('PYRUSDT', start_time = '2023-02-02',
                                    progressbar = F)

#Add the RSI
dat2 <- add_rsi(dat, time_period = "30 minutes", cutoff = 30)

#Perform backtesting
backtest(data = dat2, start_asap = F)
}
