% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covered_deviation.R
\name{covered_deviation}
\alias{covered_deviation}
\title{Calculate the covered price deviation for a specific parameter setting}
\usage{
covered_deviation(n_safety_orders, pricescale, stepscale)
}
\arguments{
\item{n_safety_orders}{The maximum number of safety orders}

\item{pricescale}{Price deviation to open safety orders (\% from initial
order)}

\item{stepscale}{With what number should the price deviation percentage used
by the last safety order be multiplied?}
}
\value{
a numeric vector of covered deviations
}
\description{
This function calculates the covered deviation (i.e. the price deviation
between the initial order and the last safety order) for a given number
of safety orders, pricescale and stepscale. This function is vectorized. So,
you can also apply it to vectors of number of safety orders, pricescales
and stepscales.
}
\examples{
covered_deviation(n_safety_orders = 2:8, pricescale = 1.5, stepscale = 1.2)
}
