% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binance_klines.R
\name{get_binance_klines}
\alias{get_binance_klines}
\title{Get kline/candlestick data from Binance}
\usage{
get_binance_klines(
  symbol = "BTCUSDT",
  start_time = Sys.time() - lubridate::ddays(3),
  end_time = Sys.time(),
  interval = c("1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d",
    "3d", "1w", "1M"),
  include_volume_data = F
)
}
\arguments{
\item{symbol}{a valid binance symbol of a trading pair}

\item{start_time, end_time}{an object of class \code{date} or \code{POSIXct}
or \code{POSIXt} or a character string that can be converted to these classes}

\item{interval}{a character string defining the time frame of the candles.}

\item{include_volume_data}{whether to include volume data or not. Default is
FALSE to save memory.}
}
\value{
a \code{data.table} with open-high-low-close values
}
\description{
This function is a wrapper for the \code{\link[=binance_klines]{binance_klines()}}
function in the \href{https://daroczig.github.io/binancer/}{\code{binancer}} package.
It improves \code{\link[=binance_klines]{binance_klines()}} by allowing to import more than 1000 candles
at once.
}
\examples{
get_binance_klines("MATICUSDT", start_time = "2022-01-01",
                    end_time = "2022-12-31", interval = "2h")
}
\seealso{
\itemize{
\item \code{\link[=get_binance_klines_from_csv]{get_binance_klines_from_csv()}} for dowloading candlestick via csv files
\item \code{\link[=get_binance_prices_from_csv]{get_binance_prices_from_csv()}} for dowloading aggTrade data via csv files
}
}
