% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binance_klines_from_csv_multi.R
\name{get_binance_klines_from_csv_multi}
\alias{get_binance_klines_from_csv_multi}
\title{Get kline/candlestick data of muliple coins from Binance by downloading and
reading csv files}
\usage{
get_binance_klines_from_csv_multi(
  coins = c("BTC", "ETH", "BNB"),
  quote_currency = "USDT",
  start_time = Sys.time() - lubridate::dmonths(2),
  end_time = Sys.time() - lubridate::ddays(2),
  interval = "1m",
  include_volume_data = F,
  quietly = F
)
}
\arguments{
\item{coins}{a vector of valid coin tickers}

\item{quote_currency}{the quote currency}

\item{start_time, end_time}{an object of class \code{date} or \code{POSIXct}
or \code{POSIXt} or a character string that can be converted to these classes}

\item{interval}{a character string specifying the time frame of the candles.
Possible values are \code{"1s"}, \code{"1m"}, \code{"3m"}, \code{"5m"}, \code{"30m"}, \code{"1h"}, \code{"2h"},
\code{"4h"}, \code{"6h"} and \code{"8h"}.}

\item{include_volume_data}{whether to include volume data or not. Default is
FALSE to save memory.}

\item{quietly}{whether to suppress progess messages}
}
\value{
a \code{data.table} with open-high-low values
}
\description{
This function downloads and reads csv files containing kline/candlestick data
of multiple coins from \url{https://data.binance.vision/} and merges the
data to a data table. To save memory, only the the columns open time,
open price, high price and low price are returned.
}
\examples{
get_binance_klines_from_csv_multi(coins = c("MATIC", "SOL", "AVAX"),
                                  start_time = '2023-01-02',
                                  end_time = '2023-01-03',
                                  interval = '1s')
}
\seealso{
\itemize{
\item \code{\link[=get_binance_klines_from_csv]{get_binance_klines_from_csv()}} for dowloading candlestick data of a
single trading pair via csv files
\item \code{\link[=get_binance_prices_from_csv]{get_binance_prices_from_csv()}} for dowloading aggTrade data via csv files
\item \code{\link[=get_binance_klines]{get_binance_klines()}} for dowloading candlestick data directly
}
}
