% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{re-exports}
\alias{re-exports}
\title{Re-exported functions for examples}
\description{
The following functions are re-exported from other packages for convenient use
in examples and vignettes.
}
\keyword{internal}
