% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_capital.R
\name{required_capital}
\alias{required_capital}
\title{Calculate the required capital for a specific parameter setting}
\usage{
required_capital(
  base_order_volume = 10,
  first_safety_order_volume = 10,
  n_safety_orders,
  volumescale
)
}
\arguments{
\item{base_order_volume}{The size of the base order (in the quote currency)}

\item{first_safety_order_volume}{The size of the first safety order (in
the quote currency)}

\item{n_safety_orders}{The maximum number of safety orders}

\item{volumescale}{With what number should the funds used by the last safety
order be multiplied?}
}
\value{
a numeric vector of required capitals
}
\description{
This function calculates the required capital for a specific parameter
setting. This function is vectorized. So, you can also apply it to vectors of
parameters.
}
\examples{
required_capital(n_safety_orders = 6:16, volumescale = 1.5)
}
